+++ changelog.md
@@ -1,5 +1,9 @@
 <h3>Changelog</h3>
 
+## 52 - Jan 28, 2015
+
+* Search with regex for function that need replacing in HTML5 player.
+
 ## 1.51 - Jan 27, 2015
 
 * Update HTML5 fix variables and check in case things have changed.